// src/app/interfaces/new-farmer.model.ts
export interface IDataNewFarmer {
  id?: string;
  name?: string;
  lastName?: string;
  gender?: number;
  documentId?: number;
  email?: string;
  telephoneNumber?: number;
  addressZipCode?: number;
  addressStreet?: string;
  addressNumber?: number;
  addressComplement?: string;
  addressDistrict?: string;
  addressCity?: string;
  addressState?: string;
  propertySize?: number;
  cultivitationType?: [];
  nourishment?: [];
  cultivationMethod?: [];
  equipmentOrMachinery?: [];
}